# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API takes a range of decimals and converts it to roman numerals",
        "title": "Roman Numeral Converter API",
        "contact": {
            "name": "Asutosh",
            "email": "asutosh.satapathy@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/api/v1",
    "paths": {
        "/convert": {
            "get": {
                "description": "Convert a comma-separated list of numbers to their corresponding Roman numeral representations.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Convert numbers to Roman numerals",
                "operationId": "convertNumbersToRoman",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated list of integers to be converted",
                        "name": "numbers",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RomanNumeral"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Convert multiple ranges of numbers to their corresponding Roman numeral representations.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Convert ranges of numbers to Roman numerals",
                "operationId": "convertRangesToRoman",
                "parameters": [
                    {
                        "description": "Array of number ranges",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RangesPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RomanNumeral"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.NumberRange": {
            "type": "object",
            "required": [
                "max",
                "min"
            ],
            "properties": {
                "max": {
                    "description": "The maximum value of the range (inclusive).",
                    "type": "integer",
                    "example": 20
                },
                "min": {
                    "description": "The minimum value of the range (inclusive).",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "models.RangesPayload": {
            "type": "object",
            "required": [
                "ranges"
            ],
            "properties": {
                "ranges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NumberRange"
                    }
                }
            }
        },
        "models.RomanNumeral": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "integer"
                },
                "roman": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}